Revert "#5-Working-on-manage-contract"
This reverts commit abbc2865ec6cc1c229ae583979608ebaf0c8dbf6.
