+++ cli/src/assets/css/app/index.scss
@@ -1,3 +1,2 @@
 // views
 @import "./views/_room";
-@import "./views/_contract";
