--- cli/src/assets/css/app/views/_contract.scss
@@ -1,3 +1,0 @@
-.contract-progress {
-
-}
