+++ cli/src/containers/Layout/AppLayout.jsx
@@ -13,7 +13,9 @@ import AboveHeader from "../Topbar/AboveHeader/index";
 import BelowHeader from "../Topbar/BelowHeader/index";
 
 import Topbar from "../Topbar/index";
+// import {footerText} from "util/config";
 import AppRoutes from "views/App/index";
+import Customizer from "containers/Customizer";
 import {connect} from "react-redux";
 import {
     NAV_STYLE_ABOVE_HEADER,
@@ -158,6 +160,7 @@ export class AppLayout extends Component {
                                 </Footer>
                             </Content>
                         </Layout>
+                        <Customizer/>
                     </Layout>
                 </IntlProvider>
             </ConfigProvider>
