--- cli/src/util/helper.js
@@ -1,7 +1,0 @@
-export function formatVND(num) {
-    return new Intl.NumberFormat("en-US", {
-        style: "currency",
-        currency: "VND",
-        currencyDisplay: "code"
-    }).format(num*1000);
-}
