+++ cli/src/views/App/Room/index.jsx
@@ -1,10 +1,9 @@
 import React, {useState, useEffect, useCallback} from "react";
-import {Row, Col, Card, Spin, Table, Button, Divider} from "antd";
+import {Card, Spin, Table, Button, Divider} from "antd";
 import withNoti from "hocs/withNoti";
 import PopConfirm from "components/App/Pop/PopConfirm";
 import api, {apiCall} from "constants/api";
-import RoomForm from "./Form";
-import RoomAssign from "./Assign";
+import RoomForm from "./RoomForm";
 
 const DEFAULT_ROOM = {
     name: "",
@@ -17,7 +16,6 @@ function Room({notify}) {
     const [price, setPrice] = useState([]);
     const [loading, setLoading] = useState(true);
     const [form, toggleForm] = useState(false);
-    const [assign, toggleAssign] = useState(false);
 
     const load = useCallback(async() => {
         try {
@@ -52,119 +50,101 @@ function Room({notify}) {
         toggleForm(true);
         setRoom(prev => ({
             ...prev, ...room,
-            price_id: room.price_id._id
+            price_id: room.price_id._id,
+            user_id: [...room.user_id]
         }));
+    }
+
+    async function hdSubmit(roomData) {
+        setLoading(true);
+        roomData.user_id = roomData.user_id.map(u => u._id);
+        try {
+            // Getting the room's user_id instead of the whole object
+            if(!roomData._id) {
+                // Create new room
+                let createdRoom = await apiCall(...api.room.create(), roomData);
+                let createdRoomData = await apiCall(...api.room.getOne(createdRoom._id));
+                let newRooms = [...rooms, createdRoomData];
+                setRooms(newRooms);
+                notify("success", "Process is completed!", "Add new room successfully!");
+            } else {
+                // Update available room data
+                let updatedRoom = await apiCall(...api.room.edit(room._id), roomData);
+                let updatedRoomData = await apiCall(...api.room.getOne(updatedRoom._id));
+                let newRooms = rooms.map(r => r._id === updatedRoom._id ? updatedRoomData : r);
+                setRooms(newRooms);
+                notify("success", "Process is completed!", "Update room successfully!");
+            }
+        } catch (e) {
+            notify("error", "Process is not completed", "The data is not submitted successfully!")
+        }
+        hdCancel();
+        setLoading(false);
     }
 
     function hdCancel() {
         setRoom(DEFAULT_ROOM);
         toggleForm(false);
-        toggleAssign(false);
-    }
-
-    function hdAssign(room) {
-        setLoading(true);
-        setRoom(prev => ({
-            ...prev, ...room,
-            user_id: [...room.user_id]
-        }));
-        toggleAssign(true);
-    }
-
-    async function refreshRoom(record, message, isNewRecord = true) {
-        if(isNewRecord) {
-            setRooms(prev => [...prev, record]);
-        } else {
-            let newRooms = rooms.map(r => r._id === record._id ? record : r);
-            setRooms(newRooms);
-        }
-        notify("success", "Process is completed!", message);
-        hdCancel();
-        setLoading(false);
-    }
-
-    function controlCols() {
-        let cols = [
-            {
-                title: "Room Name",
-                dataIndex: 'name',
-            },
-            {
-                title: "People",
-                dataIndex: 'user_id',
-                render: text => <span>{text.length} people</span>
-            },
-            {
-                title: 'Price Type',
-                dataIndex: 'price_id.type'
-            },
-            {
-                title: 'Action',
-                key: 'action',
-                render: (text, record) => room._id ? <span>None</span> : (
-                    <span>
-                        <PopConfirm
-                            title="Are you sure to delete this genre?"
-                            task={hdRemove.bind(this, record._id)}
-                            okText="Sure, remove it"
-                            cancelText="Not now"
-                        >
-                            <span className="gx-link">Delete</span>
-                        </PopConfirm>
-                        <Divider type="vertical"/>
-                        <span className="gx-link" onClick={hdEdit.bind(this, record)}>Edit</span>
-                        <Divider type="vertical"/>
-                        <span className="gx-link" onClick={hdAssign.bind(this, record)}>Assign</span>
-                    </span>
-                )
-            }
-        ];
-        return (assign || form) ? cols.filter(c => c.key !== "action") : cols;
     }
 
     return (
         <div>
-            <Row>
-                {
-                    form && <Col md={10}>
-                        <RoomForm
-                            price={price}
-                            loading={loading}
-                            setLoading={setLoading}
-                            editRoom={room}
-                            refresh={refreshRoom}
-                            hdCancel={hdCancel}
-                        />
-                    </Col>
-                }
-                <Col md={(assign || form) ? 14 : 24}>
-                    <Card title="List of available room">
-                        <Spin spinning={loading}>
+            {
+                form && <RoomForm
+                    price={price}
+                    loading={loading}
+                    setLoading={setLoading}
+                    hdSubmit={hdSubmit}
+                    editRoom={room}
+                    hdCancel={hdCancel}
+                    notify={notify}
+                />
+            }
+            <Card title="List of available room">
+                <Spin spinning={loading}>
+                    {
+                        form || <Button type="primary" onClick={() => toggleForm(true)}>Add new room information</Button>
+                    }
+                    <Table
+                        className="gx-table-responsive"
+                        dataSource={rooms}
+                        rowKey="_id"
+                        columns={[
                             {
-                                form || <Button type="primary" onClick={() => toggleForm(true)}>Add new room information</Button>
+                                title: "Room Name",
+                                dataIndex: 'name',
+                            },
+                            {
+                                title: "People",
+                                dataIndex: 'user_id',
+                                render: text => <span>{text.length} people</span>
+                            },
+                            {
+                                title: 'Price Type',
+                                dataIndex: 'price_id.type'
+                            },
+                            {
+                                title: 'Action',
+                                key: 'action',
+                                render: (text, record) => room._id ? <span>None</span> : (
+                                    <span>
+                                        <PopConfirm
+                                            title="Are you sure to delete this genre?"
+                                            task={hdRemove.bind(this, record._id)}
+                                            okText="Sure, remove it"
+                                            cancelText="Not now"
+                                        >
+                                            <span className="gx-link">Delete</span>
+                                        </PopConfirm>
+                                        <Divider type="vertical"/>
+                                        <span className="gx-link" onClick={hdEdit.bind(this, record)}>Edit</span>
+                                    </span>
+                                )
                             }
-                            <Table
-                                className="gx-table-responsive"
-                                dataSource={rooms}
-                                rowKey="_id"
-                                columns={controlCols()}
-                            />
-                        </Spin>
-                    </Card>
-                </Col>
-                {
-                    assign && <Col md={10}>
-                        <RoomAssign
-                            loading={loading}
-                            setLoading={setLoading}
-                            refresh={refreshRoom}
-                            selectedRoom={room}
-                            hdCancel={hdCancel}
-                            notify={notify}
-                        />
-                    </Col>
-                }
-            </Row>
+                        ]}
+                    />
+                </Spin>
+            </Card>
         </div>
     )
 }
