+++ package.json
@@ -38,7 +38,6 @@
         "request": "^2.88.0"
     },
     "devDependencies": {
-        "eslint": "^6.8.0",
         "expect.js": "^0.3.1",
         "mocha": "^6.2.2",
         "node-mocks-http": "^1.8.1",
